+++ .documentary/index.jsx
@@ -2,10 +2,10 @@
  * The footer for documentation.
  */
 export const footer = () => {
-  const alt = '{{ org }}'
-  const src = 'https://avatars3.githubusercontent.com/u/38815725?v=4&s=100'
-  const href = '{{ website }}'
-  const org = '{{ trademark }}'
+  const alt = 'art-deco'
+  const src = 'https://avatars2.githubusercontent.com/u/57873407?v=4&s=100'
+  const href = 'https://artd.eco'
+  const org = 'Art Deco'
   const year = new Date().getFullYear()
   return [
     (<table>