+++ example/index.js
@@ -1,7 +1,7 @@
-import myNewPackage from '../src'
+import examplePackage from '../src'
 
 (async () => {
-  const res = await myNewPackage({
+  const res = await examplePackage({
     text: 'example',
   })
   console.log(res)