+++ test/context/index.js
@@ -2,7 +2,7 @@ import { join } from 'path'
 import { debuglog } from 'util'
 import { readFileSync } from 'fs'
 
-const LOG = debuglog('my-new-package')
+const LOG = debuglog('@artdeco/example-package')
 
 /**
  * A testing context for the package.
@@ -46,16 +46,11 @@ export default class Context {
   }
 }
 
-/* bin-start */
 let BIN = 'src/BIN'
-if (process.env.ALAMODE_ENV == 'test-build') {
-  console.log('Testing build bin...')
-  BIN = 'build/bin/mnp'
-} else if (process.env.ALAMODE_ENV == 'test-compile') {
+if (process.env.ALAMODE_ENV == 'test-compile') {
   console.log('Testing compile bin...')
-  BIN = 'compile/bin/mnp'
+  BIN = 'compile/bin/example-package'
 }
-/* bin-end */
 
 /** @typedef {Object<string, Test & TestSuite4>} TestSuite */
 /** @typedef {Object<string, Test & TestSuite3>} TestSuite4 */