+++ test/spec/default.js
@@ -1,20 +1,20 @@
 import { equal } from '@zoroaster/assert'
 import Context from '../context'
-import myNewPackage from '../../src'
+import examplePackage from '../../src'
 
 /** @type {TestSuite} */
 const T = {
   context: Context,
   'is a function'() {
-    equal(typeof myNewPackage, 'function')
+    equal(typeof examplePackage, 'function')
   },
   async 'calls package without error'() {
-    await myNewPackage()
+    await examplePackage()
   },
   async 'gets a link to the fixture'({ fixture, readFile }) {
     const path = fixture`test.txt`
     const text = readFile(path)
-    const res = await myNewPackage({
+    const res = await examplePackage({
       text,
     })
     equal(res, text)